# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WA Pharmacy",
    "version": "1.0.0",
    "description": "My API Description"
  },
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Search results count",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Search results page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "John Doe",
                    "address": "123 Main St",
                    "phone": "555-1234"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "name": "Jane Doe",
                "username": "janeDoe",
                "password": "abc123",
                "role": "owner"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 2,
                  "name": "Jane Doe",
                  "address": "456 Oak St",
                  "phone": "555-5678"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get a specific user by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "John Doe",
                  "address": "123 Main St",
                  "phone": "555-1234"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific user by ID",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "name": "Updated Doe"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "Updated Doe"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/archive": {
      "put": {
        "summary": "Soft delete a specific user by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "John Doe",
                  "address": "123 Main St",
                  "phone": "555-1234"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/delete": {
      "delete": {
        "summary": "Delete a specific user by ID",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          }
        }
      }
    }
  }
}